Moved model classes into dto package
