+++ bundles/org.openhab.binding.avmfritz/src/main/java/org/openhab/binding/avmfritz/internal/dto/ButtonModel.java
@@ -10,9 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.avmfritz.internal.ahamodel;
-
-import java.math.BigDecimal;
+package org.openhab.binding.avmfritz.internal.dto;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -25,24 +23,22 @@ import javax.xml.bind.annotation.XmlType;
  * @author Christoph Weitkamp - Initial contribution
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(propOrder = { "state" })
-@XmlRootElement(name = "alert")
-public class AlertModel {
-    public static final BigDecimal ON = BigDecimal.ONE;
-    public static final BigDecimal OFF = BigDecimal.ZERO;
+@XmlType(propOrder = { "lastpressedtimestamp" })
+@XmlRootElement(name = "button")
+public class ButtonModel {
 
-    private BigDecimal state;
+    private int lastpressedtimestamp;
 
-    public BigDecimal getState() {
-        return state;
+    public int getLastpressedtimestamp() {
+        return lastpressedtimestamp;
     }
 
-    public void setState(BigDecimal state) {
-        this.state = state;
+    public void setLastpressedtimestamp(int lastpressedtimestamp) {
+        this.lastpressedtimestamp = lastpressedtimestamp;
     }
 
     @Override
     public String toString() {
-        return new StringBuilder().append("[state=").append(state).append("]").toString();
+        return new StringBuilder().append("[lastpressedtimestamp=").append(lastpressedtimestamp).append("]").toString();
     }
 }
