+++ bundles/org.openhab.binding.avmfritz/src/main/java/org/openhab/binding/avmfritz/internal/dto/templates/DeviceListModel.java
@@ -10,10 +10,14 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.avmfritz.internal.ahamodel.templates;
+package org.openhab.binding.avmfritz.internal.dto.templates;
+
+import java.util.Collections;
+import java.util.List;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
 /**
@@ -22,11 +26,21 @@ import javax.xml.bind.annotation.XmlRootElement;
  * @author Christoph Weitkamp - Initial contribution
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlRootElement(name = "applymask")
-public class ApplyMaskListModel {
+@XmlRootElement(name = "devices")
+public class DeviceListModel {
+
+    @XmlElement(name = "device")
+    private List<DeviceModel> devices;
+
+    public List<DeviceModel> getDevices() {
+        if (devices == null) {
+            devices = Collections.emptyList();
+        }
+        return devices;
+    }
 
     @Override
     public String toString() {
-        return new StringBuilder().append("[]").toString();
+        return new StringBuilder().append("[devices=").append(devices).append("]").toString();
     }
 }
